requirements.txt
django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
factory-boy==3.2.0
django-elasticsearch-dsl-drf==0.22.1
